requirements.txt
# Web Framework
django>=4.2.0           # Core backend framework [cite: 15, 17]
djangorestframework     # For building the REST API (POST/GET /expenses) [cite: 18, 19, 26]

# Data Correctness & Reliability
django-cors-headers     # To allow the UI to talk to the API if hosted separately 
django-idempotency-key  # To handle retries and prevent duplicate entries [cite: 24, 52, 74]

# Production & Deployment
gunicorn                # Production WSGI HTTP Server for deployment 
whitenoise              # To serve static files (like your app.js) in production 
python-dotenv           # To manage environment variables (SECRET_KEY, DB_URL) [cite: 13, 16]

# Optional: Real Money Handling
decimal                 # Standard library (no install needed), but ensure it is used [cite: 33, 74]